requirements.txt
Flask
Flask-Cors
torch
torchvision
Pillow
numpy
gunicorn
